[Add] Global 예외 처리 validation 추가 및 로그 출력
