+++ backend/src/main/java/com/synergy/backend/global/exception/BaseException.java
@@ -8,7 +8,18 @@ import lombok.Setter;
 
 @Getter
 @Setter
-@AllArgsConstructor
 public class BaseException extends Exception {
     private BaseResponseStatus status;
+
+    // 기본 예외 메시지 확인
+    public BaseException(BaseResponseStatus status) {
+        super(status.getMessage());
+        this.status = status;
+    }
+
+    // 특정 예외 메시지 확인 (예상치 못한 경우)
+    public BaseException(BaseResponseStatus status,Throwable cause) {
+        super(cause);
+        this.status = status;
+    }
 }
